LEGO-909: Report index info even if the table has no indexes on it
Previously, if one of the requested tables existed but had no indexes on it, then
no information at all was returned for that table.  This meant that new indexes
did not get created for it.  Making sure an empty set shows up for tables that
have no indexes will allow the new indexes to get created as expected.