JDEV-30974: Stop using finalize
This switches SQLProcessor from using finalize to using a PhantomReference.
Its reference queue cleanup / poll is done by code that is about to request
a new connection.  This should occur frequently enough as to provide a good
opportunity to catch them when it counts, but no so frequently as to prove
a performance problem.  The references are explicitly cleared when the
connection is properly closed, to the reference queue should always be empty
unless there is a problem.  This means that the reference queue poll should
normally reduce to a single volatile read.